+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 
 - Look and feel: adopt the button faces used by Emacs' customization support.
 
-- New widget to allow widget-based editing of a PostgreSQL JSON or JSONB value, which is represented
-  in Emacs Lisp as a hashtable.
+- New widget to allow widget-based editing of a PostgreSQL JSON or JSONB value.
+
+- When displaying a table, show whether row-level security is enabled for this table.
 
 
 ## [0.7] - 2024-05-23